+++ corcovado/src/sys/unix/epoll.rs
@@ -291,7 +291,7 @@ const MILLIS_PER_SEC: u64 = 1_000;
 /// million years.
 pub fn millis(duration: Duration) -> u64 {
     // Round up.
-    let millis = (duration.subsec_nanos() + NANOS_PER_MILLI - 1) / NANOS_PER_MILLI;
+    let millis = duration.subsec_nanos().div_ceil(NANOS_PER_MILLI);
     duration
         .as_secs()
         .saturating_mul(MILLIS_PER_SEC)