+++ frontends/rioterm/src/cli.rs
@@ -40,8 +40,8 @@ pub struct TerminalOptions {
     pub enable_log_file: bool,
 
     /// Start window with specified title
-    #[clap(long)]
-    pub title: Option<String>,
+    #[clap(long, name = "title-placeholder")]
+    pub title_placeholder: Option<String>,
 }
 
 impl TerminalOptions {