+++ docs/docs/releases.md
@@ -10,6 +10,20 @@ language: 'en'
 
 ## 0.2.3 (unreleased)
 
+- Shell integration.
+	- OSC 7 Escape sequences to advise the terminal of the working directory.
+	- OSC 133 Escape sequence to define Input, Output and Prompt zones.
+	- OSC 1337 Escape sequences to set user vars for tracking additional shell state.
+- Rio now allows you to configure window title through configuration via template. Possible options:
+	- `TITLE`: terminal title via OSC sequences for setting terminal title
+	- `PROGRAM`: (e.g `fish`, `zsh`, `bash`, `vim`, etc...)
+	- `ABSOLUTE_PATH`: (e.g `/Users/rapha/Documents/a/rio`)
+	<!-- - `CANONICAL_PATH`: (e.g `.../Documents/a/rio`, `~/Documents/a`) -->
+	- `COLUMNS`: current columns
+	- `LINES`: current lines
+		- So, for example if you have: `{{COLUMNS}}x{{LINES}}` would show something like `88x66`.
+- Window title is now updated regardless the Navigation Mode.
+- Performance: Background and foreground data are only retrieved if is asked (either color automation is enabled or `window.title` contains any request for it).
 - Fixed: Nix build [#853](https://github.com/raphamorim/rio/pull/853).
 - Support to `window.macos-use-shadow` (enable or disable shadow on MacOS).
 - Support to `window.windows-corner-preference` (options: `Default`, `DoNotRound`,`Round` and `RoundSmall`).
@@ -17,6 +31,7 @@ language: 'en'
 - Support to `window.windows-use-no-redirection-bitmap` (This sets `WS_EX_NOREDIRECTIONBITMAP`).
 - Support for Unicode 16 characters.
 - Support to line height.
+- Renamed `--title` to `--title-placeholder` on CLI.
 - Fixed: Deb package name 'rio' conflicts with existing one in Ubuntu [#876](https://github.com/raphamorim/rio/issues/876).
 - Fixed: Unremovable bottom padding when using line-height [#449](https://github.com/raphamorim/rio/issues/449).
 - On macOS, fixed undocumented cursors (e.g. zoom, resize, help) always appearing to be invalid and falling back to the default cursor.
@@ -30,6 +45,8 @@ language: 'en'
 - Escape sequence to move cursor forward tabs ( CSI Ps I ).
 - Always emit `1` for the first parameter when having modifiers in kitty keyboard protocol.
 - Microsoft Windows: fix the event loop not waking on accessibility requests.
+- Wayland: disable title text drawn with crossfont crate, use ab_glyph crate instead.
+- Sugarloaf: Expose wgpu.
 
 ## 0.2.2